@@ -2,11 +2,12 @@
 // versions:
 // 	protoc-gen-go v1.34.2
 // 	protoc        v4.24.3
-// source: api/articles/articles.proto
+// source: articles/articles.proto
 
 package articles
 
 import (
+	_ "google.golang.org/genproto/googleapis/api/annotations"
 	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
@@ -37,7 +38,7 @@ type CreateArticlesRequest struct {
 func (x *CreateArticlesRequest) Reset() {
 	*x = CreateArticlesRequest{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_api_articles_articles_proto_msgTypes[0]
+		mi := &file_articles_articles_proto_msgTypes[0]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -50,7 +51,7 @@ func (x *CreateArticlesRequest) String() string {
 func (*CreateArticlesRequest) ProtoMessage() {}
 
 func (x *CreateArticlesRequest) ProtoReflect() protoreflect.Message {
-	mi := &file_api_articles_articles_proto_msgTypes[0]
+	mi := &file_articles_articles_proto_msgTypes[0]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -63,7 +64,7 @@ func (x *CreateArticlesRequest) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use CreateArticlesRequest.ProtoReflect.Descriptor instead.
 func (*CreateArticlesRequest) Descriptor() ([]byte, []int) {
-	return file_api_articles_articles_proto_rawDescGZIP(), []int{0}
+	return file_articles_articles_proto_rawDescGZIP(), []int{0}
 }
 
 func (x *CreateArticlesRequest) GetTitle() string {
@@ -127,7 +128,7 @@ type CreateArticlesReply struct {
 func (x *CreateArticlesReply) Reset() {
 	*x = CreateArticlesReply{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_api_articles_articles_proto_msgTypes[1]
+		mi := &file_articles_articles_proto_msgTypes[1]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -140,7 +141,7 @@ func (x *CreateArticlesReply) String() string {
 func (*CreateArticlesReply) ProtoMessage() {}
 
 func (x *CreateArticlesReply) ProtoReflect() protoreflect.Message {
-	mi := &file_api_articles_articles_proto_msgTypes[1]
+	mi := &file_articles_articles_proto_msgTypes[1]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -153,7 +154,7 @@ func (x *CreateArticlesReply) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use CreateArticlesReply.ProtoReflect.Descriptor instead.
 func (*CreateArticlesReply) Descriptor() ([]byte, []int) {
-	return file_api_articles_articles_proto_rawDescGZIP(), []int{1}
+	return file_articles_articles_proto_rawDescGZIP(), []int{1}
 }
 
 func (x *CreateArticlesReply) GetCode() uint32 {
@@ -182,7 +183,7 @@ type GetArticlesRequest struct {
 func (x *GetArticlesRequest) Reset() {
 	*x = GetArticlesRequest{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_api_articles_articles_proto_msgTypes[2]
+		mi := &file_articles_articles_proto_msgTypes[2]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -195,7 +196,7 @@ func (x *GetArticlesRequest) String() string {
 func (*GetArticlesRequest) ProtoMessage() {}
 
 func (x *GetArticlesRequest) ProtoReflect() protoreflect.Message {
-	mi := &file_api_articles_articles_proto_msgTypes[2]
+	mi := &file_articles_articles_proto_msgTypes[2]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -208,7 +209,7 @@ func (x *GetArticlesRequest) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use GetArticlesRequest.ProtoReflect.Descriptor instead.
 func (*GetArticlesRequest) Descriptor() ([]byte, []int) {
-	return file_api_articles_articles_proto_rawDescGZIP(), []int{2}
+	return file_articles_articles_proto_rawDescGZIP(), []int{2}
 }
 
 func (x *GetArticlesRequest) GetPageSize() uint32 {
@@ -237,7 +238,7 @@ type GetArticlesReply struct {
 func (x *GetArticlesReply) Reset() {
 	*x = GetArticlesReply{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_api_articles_articles_proto_msgTypes[3]
+		mi := &file_articles_articles_proto_msgTypes[3]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -250,7 +251,7 @@ func (x *GetArticlesReply) String() string {
 func (*GetArticlesReply) ProtoMessage() {}
 
 func (x *GetArticlesReply) ProtoReflect() protoreflect.Message {
-	mi := &file_api_articles_articles_proto_msgTypes[3]
+	mi := &file_articles_articles_proto_msgTypes[3]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -263,7 +264,7 @@ func (x *GetArticlesReply) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use GetArticlesReply.ProtoReflect.Descriptor instead.
 func (*GetArticlesReply) Descriptor() ([]byte, []int) {
-	return file_api_articles_articles_proto_rawDescGZIP(), []int{3}
+	return file_articles_articles_proto_rawDescGZIP(), []int{3}
 }
 
 func (x *GetArticlesReply) GetSelectedArticles() []*GetArticlesReply_ArticleInfo {
@@ -291,7 +292,7 @@ type GetSingleArticleRequest struct {
 func (x *GetSingleArticleRequest) Reset() {
 	*x = GetSingleArticleRequest{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_api_articles_articles_proto_msgTypes[4]
+		mi := &file_articles_articles_proto_msgTypes[4]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -304,7 +305,7 @@ func (x *GetSingleArticleRequest) String() string {
 func (*GetSingleArticleRequest) ProtoMessage() {}
 
 func (x *GetSingleArticleRequest) ProtoReflect() protoreflect.Message {
-	mi := &file_api_articles_articles_proto_msgTypes[4]
+	mi := &file_articles_articles_proto_msgTypes[4]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -317,7 +318,7 @@ func (x *GetSingleArticleRequest) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use GetSingleArticleRequest.ProtoReflect.Descriptor instead.
 func (*GetSingleArticleRequest) Descriptor() ([]byte, []int) {
-	return file_api_articles_articles_proto_rawDescGZIP(), []int{4}
+	return file_articles_articles_proto_rawDescGZIP(), []int{4}
 }
 
 func (x *GetSingleArticleRequest) GetArticleID() uint32 {
@@ -340,7 +341,7 @@ type GetSingleArticleReply struct {
 func (x *GetSingleArticleReply) Reset() {
 	*x = GetSingleArticleReply{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_api_articles_articles_proto_msgTypes[5]
+		mi := &file_articles_articles_proto_msgTypes[5]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -353,7 +354,7 @@ func (x *GetSingleArticleReply) String() string {
 func (*GetSingleArticleReply) ProtoMessage() {}
 
 func (x *GetSingleArticleReply) ProtoReflect() protoreflect.Message {
-	mi := &file_api_articles_articles_proto_msgTypes[5]
+	mi := &file_articles_articles_proto_msgTypes[5]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -366,7 +367,7 @@ func (x *GetSingleArticleReply) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use GetSingleArticleReply.ProtoReflect.Descriptor instead.
 func (*GetSingleArticleReply) Descriptor() ([]byte, []int) {
-	return file_api_articles_articles_proto_rawDescGZIP(), []int{5}
+	return file_articles_articles_proto_rawDescGZIP(), []int{5}
 }
 
 func (x *GetSingleArticleReply) GetArticle() *GetSingleArticleReply_RespondMsg {
@@ -406,7 +407,7 @@ type UpdateArticlesRequest struct {
 func (x *UpdateArticlesRequest) Reset() {
 	*x = UpdateArticlesRequest{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_api_articles_articles_proto_msgTypes[6]
+		mi := &file_articles_articles_proto_msgTypes[6]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -419,7 +420,7 @@ func (x *UpdateArticlesRequest) String() string {
 func (*UpdateArticlesRequest) ProtoMessage() {}
 
 func (x *UpdateArticlesRequest) ProtoReflect() protoreflect.Message {
-	mi := &file_api_articles_articles_proto_msgTypes[6]
+	mi := &file_articles_articles_proto_msgTypes[6]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -432,7 +433,7 @@ func (x *UpdateArticlesRequest) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use UpdateArticlesRequest.ProtoReflect.Descriptor instead.
 func (*UpdateArticlesRequest) Descriptor() ([]byte, []int) {
-	return file_api_articles_articles_proto_rawDescGZIP(), []int{6}
+	return file_articles_articles_proto_rawDescGZIP(), []int{6}
 }
 
 func (x *UpdateArticlesRequest) GetArticleID() uint32 {
@@ -489,7 +490,7 @@ type UpdateArticlesReply struct {
 func (x *UpdateArticlesReply) Reset() {
 	*x = UpdateArticlesReply{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_api_articles_articles_proto_msgTypes[7]
+		mi := &file_articles_articles_proto_msgTypes[7]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -502,7 +503,7 @@ func (x *UpdateArticlesReply) String() string {
 func (*UpdateArticlesReply) ProtoMessage() {}
 
 func (x *UpdateArticlesReply) ProtoReflect() protoreflect.Message {
-	mi := &file_api_articles_articles_proto_msgTypes[7]
+	mi := &file_articles_articles_proto_msgTypes[7]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -515,7 +516,7 @@ func (x *UpdateArticlesReply) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use UpdateArticlesReply.ProtoReflect.Descriptor instead.
 func (*UpdateArticlesReply) Descriptor() ([]byte, []int) {
-	return file_api_articles_articles_proto_rawDescGZIP(), []int{7}
+	return file_articles_articles_proto_rawDescGZIP(), []int{7}
 }
 
 func (x *UpdateArticlesReply) GetCode() uint32 {
@@ -543,7 +544,7 @@ type DeleteArticlesRequest struct {
 func (x *DeleteArticlesRequest) Reset() {
 	*x = DeleteArticlesRequest{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_api_articles_articles_proto_msgTypes[8]
+		mi := &file_articles_articles_proto_msgTypes[8]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -556,7 +557,7 @@ func (x *DeleteArticlesRequest) String() string {
 func (*DeleteArticlesRequest) ProtoMessage() {}
 
 func (x *DeleteArticlesRequest) ProtoReflect() protoreflect.Message {
-	mi := &file_api_articles_articles_proto_msgTypes[8]
+	mi := &file_articles_articles_proto_msgTypes[8]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -569,7 +570,7 @@ func (x *DeleteArticlesRequest) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use DeleteArticlesRequest.ProtoReflect.Descriptor instead.
 func (*DeleteArticlesRequest) Descriptor() ([]byte, []int) {
-	return file_api_articles_articles_proto_rawDescGZIP(), []int{8}
+	return file_articles_articles_proto_rawDescGZIP(), []int{8}
 }
 
 func (x *DeleteArticlesRequest) GetArticleID() uint32 {
@@ -591,7 +592,7 @@ type DeleteArticlesReply struct {
 func (x *DeleteArticlesReply) Reset() {
 	*x = DeleteArticlesReply{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_api_articles_articles_proto_msgTypes[9]
+		mi := &file_articles_articles_proto_msgTypes[9]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -604,7 +605,7 @@ func (x *DeleteArticlesReply) String() string {
 func (*DeleteArticlesReply) ProtoMessage() {}
 
 func (x *DeleteArticlesReply) ProtoReflect() protoreflect.Message {
-	mi := &file_api_articles_articles_proto_msgTypes[9]
+	mi := &file_articles_articles_proto_msgTypes[9]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -617,7 +618,7 @@ func (x *DeleteArticlesReply) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use DeleteArticlesReply.ProtoReflect.Descriptor instead.
 func (*DeleteArticlesReply) Descriptor() ([]byte, []int) {
-	return file_api_articles_articles_proto_rawDescGZIP(), []int{9}
+	return file_articles_articles_proto_rawDescGZIP(), []int{9}
 }
 
 func (x *DeleteArticlesReply) GetCode() uint32 {
@@ -649,7 +650,7 @@ type GetArticlesReply_ArticleInfo struct {
 func (x *GetArticlesReply_ArticleInfo) Reset() {
 	*x = GetArticlesReply_ArticleInfo{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_api_articles_articles_proto_msgTypes[10]
+		mi := &file_articles_articles_proto_msgTypes[10]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -662,7 +663,7 @@ func (x *GetArticlesReply_ArticleInfo) String() string {
 func (*GetArticlesReply_ArticleInfo) ProtoMessage() {}
 
 func (x *GetArticlesReply_ArticleInfo) ProtoReflect() protoreflect.Message {
-	mi := &file_api_articles_articles_proto_msgTypes[10]
+	mi := &file_articles_articles_proto_msgTypes[10]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -675,7 +676,7 @@ func (x *GetArticlesReply_ArticleInfo) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use GetArticlesReply_ArticleInfo.ProtoReflect.Descriptor instead.
 func (*GetArticlesReply_ArticleInfo) Descriptor() ([]byte, []int) {
-	return file_api_articles_articles_proto_rawDescGZIP(), []int{3, 0}
+	return file_articles_articles_proto_rawDescGZIP(), []int{3, 0}
 }
 
 func (x *GetArticlesReply_ArticleInfo) GetUid() uint32 {
@@ -729,7 +730,7 @@ type GetSingleArticleReply_RespondMsg struct {
 func (x *GetSingleArticleReply_RespondMsg) Reset() {
 	*x = GetSingleArticleReply_RespondMsg{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_api_articles_articles_proto_msgTypes[11]
+		mi := &file_articles_articles_proto_msgTypes[11]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -742,7 +743,7 @@ func (x *GetSingleArticleReply_RespondMsg) String() string {
 func (*GetSingleArticleReply_RespondMsg) ProtoMessage() {}
 
 func (x *GetSingleArticleReply_RespondMsg) ProtoReflect() protoreflect.Message {
-	mi := &file_api_articles_articles_proto_msgTypes[11]
+	mi := &file_articles_articles_proto_msgTypes[11]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -755,7 +756,7 @@ func (x *GetSingleArticleReply_RespondMsg) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use GetSingleArticleReply_RespondMsg.ProtoReflect.Descriptor instead.
 func (*GetSingleArticleReply_RespondMsg) Descriptor() ([]byte, []int) {
-	return file_api_articles_articles_proto_rawDescGZIP(), []int{5, 0}
+	return file_articles_articles_proto_rawDescGZIP(), []int{5, 0}
 }
 
 func (x *GetSingleArticleReply_RespondMsg) GetCreatedAt() string {
@@ -800,139 +801,140 @@ func (x *GetSingleArticleReply_RespondMsg) GetPageView() uint32 {
 	return 0
 }
 
-var File_api_articles_articles_proto protoreflect.FileDescriptor
+var File_articles_articles_proto protoreflect.FileDescriptor
 
-var file_api_articles_articles_proto_rawDesc = []byte{
-	0x0a, 0x1b, 0x61, 0x70, 0x69, 0x2f, 0x61, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x2f, 0x61,
-	0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x0c, 0x61,
-	0x70, 0x69, 0x2e, 0x61, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x22, 0xad, 0x01, 0x0a, 0x15,
-	0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x52, 0x65,
-	0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x14, 0x0a, 0x05, 0x54, 0x69, 0x74, 0x6c, 0x65, 0x18, 0x01,
-	0x20, 0x01, 0x28, 0x09, 0x52, 0x05, 0x54, 0x69, 0x74, 0x6c, 0x65, 0x12, 0x10, 0x0a, 0x03, 0x43,
-	0x69, 0x64, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x03, 0x43, 0x69, 0x64, 0x12, 0x12, 0x0a,
-	0x04, 0x44, 0x65, 0x73, 0x63, 0x18, 0x03, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x44, 0x65, 0x73,
-	0x63, 0x12, 0x18, 0x0a, 0x07, 0x43, 0x6f, 0x6e, 0x74, 0x65, 0x6e, 0x74, 0x18, 0x04, 0x20, 0x01,
-	0x28, 0x09, 0x52, 0x07, 0x43, 0x6f, 0x6e, 0x74, 0x65, 0x6e, 0x74, 0x12, 0x10, 0x0a, 0x03, 0x49,
-	0x6d, 0x67, 0x18, 0x05, 0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x49, 0x6d, 0x67, 0x12, 0x1a, 0x0a,
-	0x08, 0x50, 0x61, 0x67, 0x65, 0x56, 0x69, 0x65, 0x77, 0x18, 0x06, 0x20, 0x01, 0x28, 0x0d, 0x52,
-	0x08, 0x50, 0x61, 0x67, 0x65, 0x56, 0x69, 0x65, 0x77, 0x12, 0x10, 0x0a, 0x03, 0x55, 0x69, 0x64,
-	0x18, 0x07, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x03, 0x55, 0x69, 0x64, 0x22, 0x3b, 0x0a, 0x13, 0x43,
-	0x72, 0x65, 0x61, 0x74, 0x65, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x52, 0x65, 0x70,
-	0x6c, 0x79, 0x12, 0x12, 0x0a, 0x04, 0x63, 0x6f, 0x64, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0d,
-	0x52, 0x04, 0x63, 0x6f, 0x64, 0x65, 0x12, 0x10, 0x0a, 0x03, 0x4d, 0x73, 0x67, 0x18, 0x02, 0x20,
-	0x01, 0x28, 0x09, 0x52, 0x03, 0x4d, 0x73, 0x67, 0x22, 0x4a, 0x0a, 0x12, 0x47, 0x65, 0x74, 0x41,
-	0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x1a,
-	0x0a, 0x08, 0x50, 0x61, 0x67, 0x65, 0x53, 0x69, 0x7a, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0d,
-	0x52, 0x08, 0x50, 0x61, 0x67, 0x65, 0x53, 0x69, 0x7a, 0x65, 0x12, 0x18, 0x0a, 0x07, 0x50, 0x61,
-	0x67, 0x65, 0x4e, 0x75, 0x6d, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x07, 0x50, 0x61, 0x67,
-	0x65, 0x4e, 0x75, 0x6d, 0x22, 0xef, 0x01, 0x0a, 0x10, 0x47, 0x65, 0x74, 0x41, 0x72, 0x74, 0x69,
-	0x63, 0x6c, 0x65, 0x73, 0x52, 0x65, 0x70, 0x6c, 0x79, 0x12, 0x56, 0x0a, 0x10, 0x53, 0x65, 0x6c,
-	0x65, 0x63, 0x74, 0x65, 0x64, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x18, 0x01, 0x20,
-	0x03, 0x28, 0x0b, 0x32, 0x2a, 0x2e, 0x61, 0x70, 0x69, 0x2e, 0x61, 0x72, 0x74, 0x69, 0x63, 0x6c,
-	0x65, 0x73, 0x2e, 0x47, 0x65, 0x74, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x52, 0x65,
-	0x70, 0x6c, 0x79, 0x2e, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x49, 0x6e, 0x66, 0x6f, 0x52,
-	0x10, 0x53, 0x65, 0x6c, 0x65, 0x63, 0x74, 0x65, 0x64, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65,
-	0x73, 0x12, 0x12, 0x0a, 0x04, 0x63, 0x6f, 0x64, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0d, 0x52,
-	0x04, 0x63, 0x6f, 0x64, 0x65, 0x1a, 0x6f, 0x0a, 0x0b, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65,
-	0x49, 0x6e, 0x66, 0x6f, 0x12, 0x10, 0x0a, 0x03, 0x55, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28,
-	0x0d, 0x52, 0x03, 0x55, 0x69, 0x64, 0x12, 0x14, 0x0a, 0x05, 0x54, 0x69, 0x74, 0x6c, 0x65, 0x18,
-	0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x05, 0x54, 0x69, 0x74, 0x6c, 0x65, 0x12, 0x12, 0x0a, 0x04,
-	0x4e, 0x61, 0x6d, 0x65, 0x18, 0x03, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x4e, 0x61, 0x6d, 0x65,
-	0x12, 0x10, 0x0a, 0x03, 0x49, 0x6d, 0x67, 0x18, 0x04, 0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x49,
-	0x6d, 0x67, 0x12, 0x12, 0x0a, 0x04, 0x44, 0x65, 0x73, 0x63, 0x18, 0x05, 0x20, 0x01, 0x28, 0x09,
-	0x52, 0x04, 0x44, 0x65, 0x73, 0x63, 0x22, 0x37, 0x0a, 0x17, 0x47, 0x65, 0x74, 0x53, 0x69, 0x6e,
-	0x67, 0x6c, 0x65, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73,
-	0x74, 0x12, 0x1c, 0x0a, 0x09, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x49, 0x44, 0x18, 0x01,
-	0x20, 0x01, 0x28, 0x0d, 0x52, 0x09, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x49, 0x44, 0x22,
-	0xb2, 0x02, 0x0a, 0x15, 0x47, 0x65, 0x74, 0x53, 0x69, 0x6e, 0x67, 0x6c, 0x65, 0x41, 0x72, 0x74,
-	0x69, 0x63, 0x6c, 0x65, 0x52, 0x65, 0x70, 0x6c, 0x79, 0x12, 0x48, 0x0a, 0x07, 0x41, 0x72, 0x74,
-	0x69, 0x63, 0x6c, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x2e, 0x2e, 0x61, 0x70, 0x69,
-	0x2e, 0x61, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x2e, 0x47, 0x65, 0x74, 0x53, 0x69, 0x6e,
-	0x67, 0x6c, 0x65, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x52, 0x65, 0x70, 0x6c, 0x79, 0x2e,
-	0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x64, 0x4d, 0x73, 0x67, 0x52, 0x07, 0x41, 0x72, 0x74, 0x69,
-	0x63, 0x6c, 0x65, 0x12, 0x12, 0x0a, 0x04, 0x63, 0x6f, 0x64, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28,
-	0x0d, 0x52, 0x04, 0x63, 0x6f, 0x64, 0x65, 0x12, 0x10, 0x0a, 0x03, 0x6d, 0x73, 0x67, 0x18, 0x03,
-	0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x6d, 0x73, 0x67, 0x1a, 0xa8, 0x01, 0x0a, 0x0a, 0x52, 0x65,
-	0x73, 0x70, 0x6f, 0x6e, 0x64, 0x4d, 0x73, 0x67, 0x12, 0x1c, 0x0a, 0x09, 0x43, 0x72, 0x65, 0x61,
-	0x74, 0x65, 0x64, 0x41, 0x74, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x09, 0x43, 0x72, 0x65,
-	0x61, 0x74, 0x65, 0x64, 0x41, 0x74, 0x12, 0x1c, 0x0a, 0x09, 0x55, 0x70, 0x64, 0x61, 0x74, 0x65,
-	0x64, 0x41, 0x74, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x09, 0x55, 0x70, 0x64, 0x61, 0x74,
-	0x65, 0x64, 0x41, 0x74, 0x12, 0x14, 0x0a, 0x05, 0x54, 0x69, 0x74, 0x6c, 0x65, 0x18, 0x03, 0x20,
-	0x01, 0x28, 0x09, 0x52, 0x05, 0x54, 0x69, 0x74, 0x6c, 0x65, 0x12, 0x12, 0x0a, 0x04, 0x44, 0x65,
-	0x73, 0x63, 0x18, 0x04, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x44, 0x65, 0x73, 0x63, 0x12, 0x18,
-	0x0a, 0x07, 0x43, 0x6f, 0x6e, 0x74, 0x65, 0x6e, 0x74, 0x18, 0x05, 0x20, 0x01, 0x28, 0x09, 0x52,
-	0x07, 0x43, 0x6f, 0x6e, 0x74, 0x65, 0x6e, 0x74, 0x12, 0x1a, 0x0a, 0x08, 0x50, 0x61, 0x67, 0x65,
-	0x56, 0x69, 0x65, 0x77, 0x18, 0x06, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x08, 0x50, 0x61, 0x67, 0x65,
-	0x56, 0x69, 0x65, 0x77, 0x22, 0xa9, 0x01, 0x0a, 0x15, 0x55, 0x70, 0x64, 0x61, 0x74, 0x65, 0x41,
-	0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x1c,
-	0x0a, 0x09, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x49, 0x44, 0x18, 0x01, 0x20, 0x01, 0x28,
-	0x0d, 0x52, 0x09, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x49, 0x44, 0x12, 0x14, 0x0a, 0x05,
-	0x54, 0x69, 0x74, 0x6c, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x05, 0x54, 0x69, 0x74,
-	0x6c, 0x65, 0x12, 0x10, 0x0a, 0x03, 0x43, 0x69, 0x64, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0d, 0x52,
-	0x03, 0x43, 0x69, 0x64, 0x12, 0x12, 0x0a, 0x04, 0x64, 0x65, 0x73, 0x63, 0x18, 0x04, 0x20, 0x01,
-	0x28, 0x09, 0x52, 0x04, 0x64, 0x65, 0x73, 0x63, 0x12, 0x18, 0x0a, 0x07, 0x63, 0x6f, 0x6e, 0x74,
-	0x65, 0x6e, 0x74, 0x18, 0x05, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x63, 0x6f, 0x6e, 0x74, 0x65,
-	0x6e, 0x74, 0x12, 0x1c, 0x0a, 0x09, 0x54, 0x68, 0x75, 0x6d, 0x62, 0x6e, 0x61, 0x69, 0x6c, 0x18,
-	0x06, 0x20, 0x01, 0x28, 0x09, 0x52, 0x09, 0x54, 0x68, 0x75, 0x6d, 0x62, 0x6e, 0x61, 0x69, 0x6c,
-	0x22, 0x3b, 0x0a, 0x13, 0x55, 0x70, 0x64, 0x61, 0x74, 0x65, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c,
-	0x65, 0x73, 0x52, 0x65, 0x70, 0x6c, 0x79, 0x12, 0x12, 0x0a, 0x04, 0x63, 0x6f, 0x64, 0x65, 0x18,
-	0x01, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x04, 0x63, 0x6f, 0x64, 0x65, 0x12, 0x10, 0x0a, 0x03, 0x4d,
-	0x73, 0x67, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x4d, 0x73, 0x67, 0x22, 0x35, 0x0a,
-	0x15, 0x44, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x52,
-	0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x1c, 0x0a, 0x09, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c,
-	0x65, 0x49, 0x44, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x09, 0x41, 0x72, 0x74, 0x69, 0x63,
-	0x6c, 0x65, 0x49, 0x44, 0x22, 0x3b, 0x0a, 0x13, 0x44, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x41, 0x72,
-	0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x52, 0x65, 0x70, 0x6c, 0x79, 0x12, 0x12, 0x0a, 0x04, 0x63,
-	0x6f, 0x64, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x04, 0x63, 0x6f, 0x64, 0x65, 0x12,
-	0x10, 0x0a, 0x03, 0x4d, 0x73, 0x67, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x4d, 0x73,
-	0x67, 0x32, 0xc9, 0x03, 0x0a, 0x08, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x12, 0x58,
-	0x0a, 0x0e, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73,
-	0x12, 0x23, 0x2e, 0x61, 0x70, 0x69, 0x2e, 0x61, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x2e,
-	0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x52, 0x65,
-	0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x21, 0x2e, 0x61, 0x70, 0x69, 0x2e, 0x61, 0x72, 0x74, 0x69,
-	0x63, 0x6c, 0x65, 0x73, 0x2e, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x41, 0x72, 0x74, 0x69, 0x63,
-	0x6c, 0x65, 0x73, 0x52, 0x65, 0x70, 0x6c, 0x79, 0x12, 0x58, 0x0a, 0x0e, 0x55, 0x70, 0x64, 0x61,
-	0x74, 0x65, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x12, 0x23, 0x2e, 0x61, 0x70, 0x69,
-	0x2e, 0x61, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x2e, 0x55, 0x70, 0x64, 0x61, 0x74, 0x65,
-	0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a,
-	0x21, 0x2e, 0x61, 0x70, 0x69, 0x2e, 0x61, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x2e, 0x55,
-	0x70, 0x64, 0x61, 0x74, 0x65, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x52, 0x65, 0x70,
-	0x6c, 0x79, 0x12, 0x58, 0x0a, 0x0e, 0x44, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x41, 0x72, 0x74, 0x69,
-	0x63, 0x6c, 0x65, 0x73, 0x12, 0x23, 0x2e, 0x61, 0x70, 0x69, 0x2e, 0x61, 0x72, 0x74, 0x69, 0x63,
-	0x6c, 0x65, 0x73, 0x2e, 0x44, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c,
-	0x65, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x21, 0x2e, 0x61, 0x70, 0x69, 0x2e,
-	0x61, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x2e, 0x44, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x41,
-	0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x52, 0x65, 0x70, 0x6c, 0x79, 0x12, 0x4f, 0x0a, 0x0b,
-	0x47, 0x65, 0x74, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x12, 0x20, 0x2e, 0x61, 0x70,
-	0x69, 0x2e, 0x61, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x2e, 0x47, 0x65, 0x74, 0x41, 0x72,
-	0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1e, 0x2e,
+var file_articles_articles_proto_rawDesc = []byte{
+	0x0a, 0x17, 0x61, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x2f, 0x61, 0x72, 0x74, 0x69, 0x63,
+	0x6c, 0x65, 0x73, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x0c, 0x61, 0x70, 0x69, 0x2e, 0x61,
+	0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x1a, 0x1c, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2f,
+	0x61, 0x70, 0x69, 0x2f, 0x61, 0x6e, 0x6e, 0x6f, 0x74, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x73, 0x2e,
+	0x70, 0x72, 0x6f, 0x74, 0x6f, 0x22, 0xad, 0x01, 0x0a, 0x15, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65,
+	0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12,
+	0x14, 0x0a, 0x05, 0x54, 0x69, 0x74, 0x6c, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x05,
+	0x54, 0x69, 0x74, 0x6c, 0x65, 0x12, 0x10, 0x0a, 0x03, 0x43, 0x69, 0x64, 0x18, 0x02, 0x20, 0x01,
+	0x28, 0x0d, 0x52, 0x03, 0x43, 0x69, 0x64, 0x12, 0x12, 0x0a, 0x04, 0x44, 0x65, 0x73, 0x63, 0x18,
+	0x03, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x44, 0x65, 0x73, 0x63, 0x12, 0x18, 0x0a, 0x07, 0x43,
+	0x6f, 0x6e, 0x74, 0x65, 0x6e, 0x74, 0x18, 0x04, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x43, 0x6f,
+	0x6e, 0x74, 0x65, 0x6e, 0x74, 0x12, 0x10, 0x0a, 0x03, 0x49, 0x6d, 0x67, 0x18, 0x05, 0x20, 0x01,
+	0x28, 0x09, 0x52, 0x03, 0x49, 0x6d, 0x67, 0x12, 0x1a, 0x0a, 0x08, 0x50, 0x61, 0x67, 0x65, 0x56,
+	0x69, 0x65, 0x77, 0x18, 0x06, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x08, 0x50, 0x61, 0x67, 0x65, 0x56,
+	0x69, 0x65, 0x77, 0x12, 0x10, 0x0a, 0x03, 0x55, 0x69, 0x64, 0x18, 0x07, 0x20, 0x01, 0x28, 0x0d,
+	0x52, 0x03, 0x55, 0x69, 0x64, 0x22, 0x3b, 0x0a, 0x13, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x41,
+	0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x52, 0x65, 0x70, 0x6c, 0x79, 0x12, 0x12, 0x0a, 0x04,
+	0x63, 0x6f, 0x64, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x04, 0x63, 0x6f, 0x64, 0x65,
+	0x12, 0x10, 0x0a, 0x03, 0x4d, 0x73, 0x67, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x4d,
+	0x73, 0x67, 0x22, 0x4a, 0x0a, 0x12, 0x47, 0x65, 0x74, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65,
+	0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x1a, 0x0a, 0x08, 0x50, 0x61, 0x67, 0x65,
+	0x53, 0x69, 0x7a, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x08, 0x50, 0x61, 0x67, 0x65,
+	0x53, 0x69, 0x7a, 0x65, 0x12, 0x18, 0x0a, 0x07, 0x50, 0x61, 0x67, 0x65, 0x4e, 0x75, 0x6d, 0x18,
+	0x02, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x07, 0x50, 0x61, 0x67, 0x65, 0x4e, 0x75, 0x6d, 0x22, 0xef,
+	0x01, 0x0a, 0x10, 0x47, 0x65, 0x74, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x52, 0x65,
+	0x70, 0x6c, 0x79, 0x12, 0x56, 0x0a, 0x10, 0x53, 0x65, 0x6c, 0x65, 0x63, 0x74, 0x65, 0x64, 0x41,
+	0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x2a, 0x2e,
 	0x61, 0x70, 0x69, 0x2e, 0x61, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x2e, 0x47, 0x65, 0x74,
-	0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x52, 0x65, 0x70, 0x6c, 0x79, 0x12, 0x5e, 0x0a,
-	0x10, 0x47, 0x65, 0x74, 0x53, 0x69, 0x6e, 0x67, 0x6c, 0x65, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c,
-	0x65, 0x12, 0x25, 0x2e, 0x61, 0x70, 0x69, 0x2e, 0x61, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73,
-	0x2e, 0x47, 0x65, 0x74, 0x53, 0x69, 0x6e, 0x67, 0x6c, 0x65, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c,
-	0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x23, 0x2e, 0x61, 0x70, 0x69, 0x2e, 0x61,
-	0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x2e, 0x47, 0x65, 0x74, 0x53, 0x69, 0x6e, 0x67, 0x6c,
-	0x65, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x52, 0x65, 0x70, 0x6c, 0x79, 0x42, 0x1c, 0x5a,
-	0x1a, 0x62, 0x6c, 0x6f, 0x67, 0x2f, 0x61, 0x70, 0x69, 0x2f, 0x61, 0x72, 0x74, 0x69, 0x63, 0x6c,
-	0x65, 0x73, 0x3b, 0x61, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x62, 0x06, 0x70, 0x72, 0x6f,
-	0x74, 0x6f, 0x33,
+	0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x52, 0x65, 0x70, 0x6c, 0x79, 0x2e, 0x41, 0x72,
+	0x74, 0x69, 0x63, 0x6c, 0x65, 0x49, 0x6e, 0x66, 0x6f, 0x52, 0x10, 0x53, 0x65, 0x6c, 0x65, 0x63,
+	0x74, 0x65, 0x64, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x12, 0x12, 0x0a, 0x04, 0x63,
+	0x6f, 0x64, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x04, 0x63, 0x6f, 0x64, 0x65, 0x1a,
+	0x6f, 0x0a, 0x0b, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x49, 0x6e, 0x66, 0x6f, 0x12, 0x10,
+	0x0a, 0x03, 0x55, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x03, 0x55, 0x69, 0x64,
+	0x12, 0x14, 0x0a, 0x05, 0x54, 0x69, 0x74, 0x6c, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52,
+	0x05, 0x54, 0x69, 0x74, 0x6c, 0x65, 0x12, 0x12, 0x0a, 0x04, 0x4e, 0x61, 0x6d, 0x65, 0x18, 0x03,
+	0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x4e, 0x61, 0x6d, 0x65, 0x12, 0x10, 0x0a, 0x03, 0x49, 0x6d,
+	0x67, 0x18, 0x04, 0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x49, 0x6d, 0x67, 0x12, 0x12, 0x0a, 0x04,
+	0x44, 0x65, 0x73, 0x63, 0x18, 0x05, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x44, 0x65, 0x73, 0x63,
+	0x22, 0x37, 0x0a, 0x17, 0x47, 0x65, 0x74, 0x53, 0x69, 0x6e, 0x67, 0x6c, 0x65, 0x41, 0x72, 0x74,
+	0x69, 0x63, 0x6c, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x1c, 0x0a, 0x09, 0x41,
+	0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x49, 0x44, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x09,
+	0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x49, 0x44, 0x22, 0xb2, 0x02, 0x0a, 0x15, 0x47, 0x65,
+	0x74, 0x53, 0x69, 0x6e, 0x67, 0x6c, 0x65, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x52, 0x65,
+	0x70, 0x6c, 0x79, 0x12, 0x48, 0x0a, 0x07, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x18, 0x01,
+	0x20, 0x01, 0x28, 0x0b, 0x32, 0x2e, 0x2e, 0x61, 0x70, 0x69, 0x2e, 0x61, 0x72, 0x74, 0x69, 0x63,
+	0x6c, 0x65, 0x73, 0x2e, 0x47, 0x65, 0x74, 0x53, 0x69, 0x6e, 0x67, 0x6c, 0x65, 0x41, 0x72, 0x74,
+	0x69, 0x63, 0x6c, 0x65, 0x52, 0x65, 0x70, 0x6c, 0x79, 0x2e, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e,
+	0x64, 0x4d, 0x73, 0x67, 0x52, 0x07, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x12, 0x12, 0x0a,
+	0x04, 0x63, 0x6f, 0x64, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x04, 0x63, 0x6f, 0x64,
+	0x65, 0x12, 0x10, 0x0a, 0x03, 0x6d, 0x73, 0x67, 0x18, 0x03, 0x20, 0x01, 0x28, 0x09, 0x52, 0x03,
+	0x6d, 0x73, 0x67, 0x1a, 0xa8, 0x01, 0x0a, 0x0a, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x64, 0x4d,
+	0x73, 0x67, 0x12, 0x1c, 0x0a, 0x09, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x64, 0x41, 0x74, 0x18,
+	0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x09, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x64, 0x41, 0x74,
+	0x12, 0x1c, 0x0a, 0x09, 0x55, 0x70, 0x64, 0x61, 0x74, 0x65, 0x64, 0x41, 0x74, 0x18, 0x02, 0x20,
+	0x01, 0x28, 0x09, 0x52, 0x09, 0x55, 0x70, 0x64, 0x61, 0x74, 0x65, 0x64, 0x41, 0x74, 0x12, 0x14,
+	0x0a, 0x05, 0x54, 0x69, 0x74, 0x6c, 0x65, 0x18, 0x03, 0x20, 0x01, 0x28, 0x09, 0x52, 0x05, 0x54,
+	0x69, 0x74, 0x6c, 0x65, 0x12, 0x12, 0x0a, 0x04, 0x44, 0x65, 0x73, 0x63, 0x18, 0x04, 0x20, 0x01,
+	0x28, 0x09, 0x52, 0x04, 0x44, 0x65, 0x73, 0x63, 0x12, 0x18, 0x0a, 0x07, 0x43, 0x6f, 0x6e, 0x74,
+	0x65, 0x6e, 0x74, 0x18, 0x05, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x43, 0x6f, 0x6e, 0x74, 0x65,
+	0x6e, 0x74, 0x12, 0x1a, 0x0a, 0x08, 0x50, 0x61, 0x67, 0x65, 0x56, 0x69, 0x65, 0x77, 0x18, 0x06,
+	0x20, 0x01, 0x28, 0x0d, 0x52, 0x08, 0x50, 0x61, 0x67, 0x65, 0x56, 0x69, 0x65, 0x77, 0x22, 0xa9,
+	0x01, 0x0a, 0x15, 0x55, 0x70, 0x64, 0x61, 0x74, 0x65, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65,
+	0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x1c, 0x0a, 0x09, 0x41, 0x72, 0x74, 0x69,
+	0x63, 0x6c, 0x65, 0x49, 0x44, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x09, 0x41, 0x72, 0x74,
+	0x69, 0x63, 0x6c, 0x65, 0x49, 0x44, 0x12, 0x14, 0x0a, 0x05, 0x54, 0x69, 0x74, 0x6c, 0x65, 0x18,
+	0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x05, 0x54, 0x69, 0x74, 0x6c, 0x65, 0x12, 0x10, 0x0a, 0x03,
+	0x43, 0x69, 0x64, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x03, 0x43, 0x69, 0x64, 0x12, 0x12,
+	0x0a, 0x04, 0x64, 0x65, 0x73, 0x63, 0x18, 0x04, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x64, 0x65,
+	0x73, 0x63, 0x12, 0x18, 0x0a, 0x07, 0x63, 0x6f, 0x6e, 0x74, 0x65, 0x6e, 0x74, 0x18, 0x05, 0x20,
+	0x01, 0x28, 0x09, 0x52, 0x07, 0x63, 0x6f, 0x6e, 0x74, 0x65, 0x6e, 0x74, 0x12, 0x1c, 0x0a, 0x09,
+	0x54, 0x68, 0x75, 0x6d, 0x62, 0x6e, 0x61, 0x69, 0x6c, 0x18, 0x06, 0x20, 0x01, 0x28, 0x09, 0x52,
+	0x09, 0x54, 0x68, 0x75, 0x6d, 0x62, 0x6e, 0x61, 0x69, 0x6c, 0x22, 0x3b, 0x0a, 0x13, 0x55, 0x70,
+	0x64, 0x61, 0x74, 0x65, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x52, 0x65, 0x70, 0x6c,
+	0x79, 0x12, 0x12, 0x0a, 0x04, 0x63, 0x6f, 0x64, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0d, 0x52,
+	0x04, 0x63, 0x6f, 0x64, 0x65, 0x12, 0x10, 0x0a, 0x03, 0x4d, 0x73, 0x67, 0x18, 0x02, 0x20, 0x01,
+	0x28, 0x09, 0x52, 0x03, 0x4d, 0x73, 0x67, 0x22, 0x35, 0x0a, 0x15, 0x44, 0x65, 0x6c, 0x65, 0x74,
+	0x65, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74,
+	0x12, 0x1c, 0x0a, 0x09, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x49, 0x44, 0x18, 0x01, 0x20,
+	0x01, 0x28, 0x0d, 0x52, 0x09, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x49, 0x44, 0x22, 0x3b,
+	0x0a, 0x13, 0x44, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73,
+	0x52, 0x65, 0x70, 0x6c, 0x79, 0x12, 0x12, 0x0a, 0x04, 0x63, 0x6f, 0x64, 0x65, 0x18, 0x01, 0x20,
+	0x01, 0x28, 0x0d, 0x52, 0x04, 0x63, 0x6f, 0x64, 0x65, 0x12, 0x10, 0x0a, 0x03, 0x4d, 0x73, 0x67,
+	0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x4d, 0x73, 0x67, 0x32, 0xc9, 0x03, 0x0a, 0x08,
+	0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x12, 0x58, 0x0a, 0x0e, 0x43, 0x72, 0x65, 0x61,
+	0x74, 0x65, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x12, 0x23, 0x2e, 0x61, 0x70, 0x69,
+	0x2e, 0x61, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x2e, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65,
+	0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a,
+	0x21, 0x2e, 0x61, 0x70, 0x69, 0x2e, 0x61, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x2e, 0x43,
+	0x72, 0x65, 0x61, 0x74, 0x65, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x52, 0x65, 0x70,
+	0x6c, 0x79, 0x12, 0x58, 0x0a, 0x0e, 0x55, 0x70, 0x64, 0x61, 0x74, 0x65, 0x41, 0x72, 0x74, 0x69,
+	0x63, 0x6c, 0x65, 0x73, 0x12, 0x23, 0x2e, 0x61, 0x70, 0x69, 0x2e, 0x61, 0x72, 0x74, 0x69, 0x63,
+	0x6c, 0x65, 0x73, 0x2e, 0x55, 0x70, 0x64, 0x61, 0x74, 0x65, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c,
+	0x65, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x21, 0x2e, 0x61, 0x70, 0x69, 0x2e,
+	0x61, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x2e, 0x55, 0x70, 0x64, 0x61, 0x74, 0x65, 0x41,
+	0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x52, 0x65, 0x70, 0x6c, 0x79, 0x12, 0x58, 0x0a, 0x0e,
+	0x44, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x12, 0x23,
+	0x2e, 0x61, 0x70, 0x69, 0x2e, 0x61, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x2e, 0x44, 0x65,
+	0x6c, 0x65, 0x74, 0x65, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x52, 0x65, 0x71, 0x75,
+	0x65, 0x73, 0x74, 0x1a, 0x21, 0x2e, 0x61, 0x70, 0x69, 0x2e, 0x61, 0x72, 0x74, 0x69, 0x63, 0x6c,
+	0x65, 0x73, 0x2e, 0x44, 0x65, 0x6c, 0x65, 0x74, 0x65, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65,
+	0x73, 0x52, 0x65, 0x70, 0x6c, 0x79, 0x12, 0x4f, 0x0a, 0x0b, 0x47, 0x65, 0x74, 0x41, 0x72, 0x74,
+	0x69, 0x63, 0x6c, 0x65, 0x73, 0x12, 0x20, 0x2e, 0x61, 0x70, 0x69, 0x2e, 0x61, 0x72, 0x74, 0x69,
+	0x63, 0x6c, 0x65, 0x73, 0x2e, 0x47, 0x65, 0x74, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73,
+	0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1e, 0x2e, 0x61, 0x70, 0x69, 0x2e, 0x61, 0x72,
+	0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x2e, 0x47, 0x65, 0x74, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c,
+	0x65, 0x73, 0x52, 0x65, 0x70, 0x6c, 0x79, 0x12, 0x5e, 0x0a, 0x10, 0x47, 0x65, 0x74, 0x53, 0x69,
+	0x6e, 0x67, 0x6c, 0x65, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x12, 0x25, 0x2e, 0x61, 0x70,
+	0x69, 0x2e, 0x61, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x2e, 0x47, 0x65, 0x74, 0x53, 0x69,
+	0x6e, 0x67, 0x6c, 0x65, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65,
+	0x73, 0x74, 0x1a, 0x23, 0x2e, 0x61, 0x70, 0x69, 0x2e, 0x61, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65,
+	0x73, 0x2e, 0x47, 0x65, 0x74, 0x53, 0x69, 0x6e, 0x67, 0x6c, 0x65, 0x41, 0x72, 0x74, 0x69, 0x63,
+	0x6c, 0x65, 0x52, 0x65, 0x70, 0x6c, 0x79, 0x42, 0x1c, 0x5a, 0x1a, 0x62, 0x6c, 0x6f, 0x67, 0x2f,
+	0x61, 0x70, 0x69, 0x2f, 0x61, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x3b, 0x61, 0x72, 0x74,
+	0x69, 0x63, 0x6c, 0x65, 0x73, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
 }
 
 var (
-	file_api_articles_articles_proto_rawDescOnce sync.Once
-	file_api_articles_articles_proto_rawDescData = file_api_articles_articles_proto_rawDesc
+	file_articles_articles_proto_rawDescOnce sync.Once
+	file_articles_articles_proto_rawDescData = file_articles_articles_proto_rawDesc
 )
 
-func file_api_articles_articles_proto_rawDescGZIP() []byte {
-	file_api_articles_articles_proto_rawDescOnce.Do(func() {
-		file_api_articles_articles_proto_rawDescData = protoimpl.X.CompressGZIP(file_api_articles_articles_proto_rawDescData)
+func file_articles_articles_proto_rawDescGZIP() []byte {
+	file_articles_articles_proto_rawDescOnce.Do(func() {
+		file_articles_articles_proto_rawDescData = protoimpl.X.CompressGZIP(file_articles_articles_proto_rawDescData)
 	})
-	return file_api_articles_articles_proto_rawDescData
+	return file_articles_articles_proto_rawDescData
 }
 
-var file_api_articles_articles_proto_msgTypes = make([]protoimpl.MessageInfo, 12)
-var file_api_articles_articles_proto_goTypes = []any{
+var file_articles_articles_proto_msgTypes = make([]protoimpl.MessageInfo, 12)
+var file_articles_articles_proto_goTypes = []any{
 	(*CreateArticlesRequest)(nil),            // 0: api.articles.CreateArticlesRequest
 	(*CreateArticlesReply)(nil),              // 1: api.articles.CreateArticlesReply
 	(*GetArticlesRequest)(nil),               // 2: api.articles.GetArticlesRequest
@@ -946,7 +948,7 @@ var file_api_articles_articles_proto_goTypes = []any{
 	(*GetArticlesReply_ArticleInfo)(nil),     // 10: api.articles.GetArticlesReply.ArticleInfo
 	(*GetSingleArticleReply_RespondMsg)(nil), // 11: api.articles.GetSingleArticleReply.RespondMsg
 }
-var file_api_articles_articles_proto_depIdxs = []int32{
+var file_articles_articles_proto_depIdxs = []int32{
 	10, // 0: api.articles.GetArticlesReply.SelectedArticles:type_name -> api.articles.GetArticlesReply.ArticleInfo
 	11, // 1: api.articles.GetSingleArticleReply.Article:type_name -> api.articles.GetSingleArticleReply.RespondMsg
 	0,  // 2: api.articles.Articles.CreateArticles:input_type -> api.articles.CreateArticlesRequest
@@ -966,13 +968,13 @@ var file_api_articles_articles_proto_depIdxs = []int32{
 	0,  // [0:2] is the sub-list for field type_name
 }
 
-func init() { file_api_articles_articles_proto_init() }
-func file_api_articles_articles_proto_init() {
-	if File_api_articles_articles_proto != nil {
+func init() { file_articles_articles_proto_init() }
+func file_articles_articles_proto_init() {
+	if File_articles_articles_proto != nil {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_api_articles_articles_proto_msgTypes[0].Exporter = func(v any, i int) any {
+		file_articles_articles_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*CreateArticlesRequest); i {
 			case 0:
 				return &v.state
@@ -984,7 +986,7 @@ func file_api_articles_articles_proto_init() {
 				return nil
 			}
 		}
-		file_api_articles_articles_proto_msgTypes[1].Exporter = func(v any, i int) any {
+		file_articles_articles_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*CreateArticlesReply); i {
 			case 0:
 				return &v.state
@@ -996,7 +998,7 @@ func file_api_articles_articles_proto_init() {
 				return nil
 			}
 		}
-		file_api_articles_articles_proto_msgTypes[2].Exporter = func(v any, i int) any {
+		file_articles_articles_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*GetArticlesRequest); i {
 			case 0:
 				return &v.state
@@ -1008,7 +1010,7 @@ func file_api_articles_articles_proto_init() {
 				return nil
 			}
 		}
-		file_api_articles_articles_proto_msgTypes[3].Exporter = func(v any, i int) any {
+		file_articles_articles_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*GetArticlesReply); i {
 			case 0:
 				return &v.state
@@ -1020,7 +1022,7 @@ func file_api_articles_articles_proto_init() {
 				return nil
 			}
 		}
-		file_api_articles_articles_proto_msgTypes[4].Exporter = func(v any, i int) any {
+		file_articles_articles_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*GetSingleArticleRequest); i {
 			case 0:
 				return &v.state
@@ -1032,7 +1034,7 @@ func file_api_articles_articles_proto_init() {
 				return nil
 			}
 		}
-		file_api_articles_articles_proto_msgTypes[5].Exporter = func(v any, i int) any {
+		file_articles_articles_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*GetSingleArticleReply); i {
 			case 0:
 				return &v.state
@@ -1044,7 +1046,7 @@ func file_api_articles_articles_proto_init() {
 				return nil
 			}
 		}
-		file_api_articles_articles_proto_msgTypes[6].Exporter = func(v any, i int) any {
+		file_articles_articles_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*UpdateArticlesRequest); i {
 			case 0:
 				return &v.state
@@ -1056,7 +1058,7 @@ func file_api_articles_articles_proto_init() {
 				return nil
 			}
 		}
-		file_api_articles_articles_proto_msgTypes[7].Exporter = func(v any, i int) any {
+		file_articles_articles_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*UpdateArticlesReply); i {
 			case 0:
 				return &v.state
@@ -1068,7 +1070,7 @@ func file_api_articles_articles_proto_init() {
 				return nil
 			}
 		}
-		file_api_articles_articles_proto_msgTypes[8].Exporter = func(v any, i int) any {
+		file_articles_articles_proto_msgTypes[8].Exporter = func(v any, i int) any {
 			switch v := v.(*DeleteArticlesRequest); i {
 			case 0:
 				return &v.state
@@ -1080,7 +1082,7 @@ func file_api_articles_articles_proto_init() {
 				return nil
 			}
 		}
-		file_api_articles_articles_proto_msgTypes[9].Exporter = func(v any, i int) any {
+		file_articles_articles_proto_msgTypes[9].Exporter = func(v any, i int) any {
 			switch v := v.(*DeleteArticlesReply); i {
 			case 0:
 				return &v.state
@@ -1092,7 +1094,7 @@ func file_api_articles_articles_proto_init() {
 				return nil
 			}
 		}
-		file_api_articles_articles_proto_msgTypes[10].Exporter = func(v any, i int) any {
+		file_articles_articles_proto_msgTypes[10].Exporter = func(v any, i int) any {
 			switch v := v.(*GetArticlesReply_ArticleInfo); i {
 			case 0:
 				return &v.state
@@ -1104,7 +1106,7 @@ func file_api_articles_articles_proto_init() {
 				return nil
 			}
 		}
-		file_api_articles_articles_proto_msgTypes[11].Exporter = func(v any, i int) any {
+		file_articles_articles_proto_msgTypes[11].Exporter = func(v any, i int) any {
 			switch v := v.(*GetSingleArticleReply_RespondMsg); i {
 			case 0:
 				return &v.state
@@ -1121,18 +1123,18 @@ func file_api_articles_articles_proto_init() {
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_api_articles_articles_proto_rawDesc,
+			RawDescriptor: file_articles_articles_proto_rawDesc,
 			NumEnums:      0,
 			NumMessages:   12,
 			NumExtensions: 0,
 			NumServices:   1,
 		},
-		GoTypes:           file_api_articles_articles_proto_goTypes,
-		DependencyIndexes: file_api_articles_articles_proto_depIdxs,
-		MessageInfos:      file_api_articles_articles_proto_msgTypes,
+		GoTypes:           file_articles_articles_proto_goTypes,
+		DependencyIndexes: file_articles_articles_proto_depIdxs,
+		MessageInfos:      file_articles_articles_proto_msgTypes,
 	}.Build()
-	File_api_articles_articles_proto = out.File
-	file_api_articles_articles_proto_rawDesc = nil
-	file_api_articles_articles_proto_goTypes = nil
-	file_api_articles_articles_proto_depIdxs = nil
+	File_articles_articles_proto = out.File
+	file_articles_articles_proto_rawDesc = nil
+	file_articles_articles_proto_goTypes = nil
+	file_articles_articles_proto_depIdxs = nil
 }
