@@ -2,7 +2,7 @@
 // versions:
 // 	protoc-gen-go v1.34.2
 // 	protoc        v4.24.3
-// source: api/articles/error_def.proto
+// source: articles/error_def.proto
 
 package articles
 
@@ -54,11 +54,11 @@ func (x ArticleErrorReason) String() string {
 }
 
 func (ArticleErrorReason) Descriptor() protoreflect.EnumDescriptor {
-	return file_api_articles_error_def_proto_enumTypes[0].Descriptor()
+	return file_articles_error_def_proto_enumTypes[0].Descriptor()
 }
 
 func (ArticleErrorReason) Type() protoreflect.EnumType {
-	return &file_api_articles_error_def_proto_enumTypes[0]
+	return &file_articles_error_def_proto_enumTypes[0]
 }
 
 func (x ArticleErrorReason) Number() protoreflect.EnumNumber {
@@ -67,45 +67,45 @@ func (x ArticleErrorReason) Number() protoreflect.EnumNumber {
 
 // Deprecated: Use ArticleErrorReason.Descriptor instead.
 func (ArticleErrorReason) EnumDescriptor() ([]byte, []int) {
-	return file_api_articles_error_def_proto_rawDescGZIP(), []int{0}
+	return file_articles_error_def_proto_rawDescGZIP(), []int{0}
 }
 
-var File_api_articles_error_def_proto protoreflect.FileDescriptor
+var File_articles_error_def_proto protoreflect.FileDescriptor
 
-var file_api_articles_error_def_proto_rawDesc = []byte{
-	0x0a, 0x1c, 0x61, 0x70, 0x69, 0x2f, 0x61, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x2f, 0x65,
-	0x72, 0x72, 0x6f, 0x72, 0x5f, 0x64, 0x65, 0x66, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x0c,
-	0x61, 0x70, 0x69, 0x2e, 0x61, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x1a, 0x13, 0x65, 0x72,
-	0x72, 0x6f, 0x72, 0x73, 0x2f, 0x65, 0x72, 0x72, 0x6f, 0x72, 0x73, 0x2e, 0x70, 0x72, 0x6f, 0x74,
-	0x6f, 0x2a, 0x73, 0x0a, 0x12, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x45, 0x72, 0x72, 0x6f,
-	0x72, 0x52, 0x65, 0x61, 0x73, 0x6f, 0x6e, 0x12, 0x1f, 0x0a, 0x15, 0x45, 0x52, 0x52, 0x5f, 0x41,
-	0x52, 0x54, 0x49, 0x43, 0x4c, 0x45, 0x5f, 0x4e, 0x4f, 0x54, 0x5f, 0x45, 0x58, 0x49, 0x53, 0x54,
-	0x10, 0x00, 0x1a, 0x04, 0xa8, 0x45, 0x94, 0x03, 0x12, 0x29, 0x0a, 0x1f, 0x45, 0x52, 0x52, 0x5f,
-	0x41, 0x52, 0x54, 0x49, 0x43, 0x4c, 0x45, 0x5f, 0x49, 0x4e, 0x56, 0x41, 0x4c, 0x49, 0x44, 0x5f,
-	0x49, 0x44, 0x5f, 0x4f, 0x52, 0x5f, 0x54, 0x49, 0x54, 0x4c, 0x45, 0x10, 0x01, 0x1a, 0x04, 0xa8,
-	0x45, 0x90, 0x03, 0x12, 0x11, 0x0a, 0x07, 0x53, 0x55, 0x43, 0x43, 0x45, 0x53, 0x53, 0x10, 0x02,
-	0x1a, 0x04, 0xa8, 0x45, 0xc8, 0x01, 0x42, 0x1c, 0x5a, 0x1a, 0x62, 0x6c, 0x6f, 0x67, 0x2f, 0x61,
-	0x70, 0x69, 0x2f, 0x61, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x3b, 0x61, 0x72, 0x74, 0x69,
-	0x63, 0x6c, 0x65, 0x73, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
+var file_articles_error_def_proto_rawDesc = []byte{
+	0x0a, 0x18, 0x61, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x2f, 0x65, 0x72, 0x72, 0x6f, 0x72,
+	0x5f, 0x64, 0x65, 0x66, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x0c, 0x61, 0x70, 0x69, 0x2e,
+	0x61, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x1a, 0x13, 0x65, 0x72, 0x72, 0x6f, 0x72, 0x73,
+	0x2f, 0x65, 0x72, 0x72, 0x6f, 0x72, 0x73, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2a, 0x73, 0x0a,
+	0x12, 0x41, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x45, 0x72, 0x72, 0x6f, 0x72, 0x52, 0x65, 0x61,
+	0x73, 0x6f, 0x6e, 0x12, 0x1f, 0x0a, 0x15, 0x45, 0x52, 0x52, 0x5f, 0x41, 0x52, 0x54, 0x49, 0x43,
+	0x4c, 0x45, 0x5f, 0x4e, 0x4f, 0x54, 0x5f, 0x45, 0x58, 0x49, 0x53, 0x54, 0x10, 0x00, 0x1a, 0x04,
+	0xa8, 0x45, 0x94, 0x03, 0x12, 0x29, 0x0a, 0x1f, 0x45, 0x52, 0x52, 0x5f, 0x41, 0x52, 0x54, 0x49,
+	0x43, 0x4c, 0x45, 0x5f, 0x49, 0x4e, 0x56, 0x41, 0x4c, 0x49, 0x44, 0x5f, 0x49, 0x44, 0x5f, 0x4f,
+	0x52, 0x5f, 0x54, 0x49, 0x54, 0x4c, 0x45, 0x10, 0x01, 0x1a, 0x04, 0xa8, 0x45, 0x90, 0x03, 0x12,
+	0x11, 0x0a, 0x07, 0x53, 0x55, 0x43, 0x43, 0x45, 0x53, 0x53, 0x10, 0x02, 0x1a, 0x04, 0xa8, 0x45,
+	0xc8, 0x01, 0x42, 0x1c, 0x5a, 0x1a, 0x62, 0x6c, 0x6f, 0x67, 0x2f, 0x61, 0x70, 0x69, 0x2f, 0x61,
+	0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73, 0x3b, 0x61, 0x72, 0x74, 0x69, 0x63, 0x6c, 0x65, 0x73,
+	0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
 }
 
 var (
-	file_api_articles_error_def_proto_rawDescOnce sync.Once
-	file_api_articles_error_def_proto_rawDescData = file_api_articles_error_def_proto_rawDesc
+	file_articles_error_def_proto_rawDescOnce sync.Once
+	file_articles_error_def_proto_rawDescData = file_articles_error_def_proto_rawDesc
 )
 
-func file_api_articles_error_def_proto_rawDescGZIP() []byte {
-	file_api_articles_error_def_proto_rawDescOnce.Do(func() {
-		file_api_articles_error_def_proto_rawDescData = protoimpl.X.CompressGZIP(file_api_articles_error_def_proto_rawDescData)
+func file_articles_error_def_proto_rawDescGZIP() []byte {
+	file_articles_error_def_proto_rawDescOnce.Do(func() {
+		file_articles_error_def_proto_rawDescData = protoimpl.X.CompressGZIP(file_articles_error_def_proto_rawDescData)
 	})
-	return file_api_articles_error_def_proto_rawDescData
+	return file_articles_error_def_proto_rawDescData
 }
 
-var file_api_articles_error_def_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
-var file_api_articles_error_def_proto_goTypes = []any{
+var file_articles_error_def_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
+var file_articles_error_def_proto_goTypes = []any{
 	(ArticleErrorReason)(0), // 0: api.articles.ArticleErrorReason
 }
-var file_api_articles_error_def_proto_depIdxs = []int32{
+var file_articles_error_def_proto_depIdxs = []int32{
 	0, // [0:0] is the sub-list for method output_type
 	0, // [0:0] is the sub-list for method input_type
 	0, // [0:0] is the sub-list for extension type_name
@@ -113,27 +113,27 @@ var file_api_articles_error_def_proto_depIdxs = []int32{
 	0, // [0:0] is the sub-list for field type_name
 }
 
-func init() { file_api_articles_error_def_proto_init() }
-func file_api_articles_error_def_proto_init() {
-	if File_api_articles_error_def_proto != nil {
+func init() { file_articles_error_def_proto_init() }
+func file_articles_error_def_proto_init() {
+	if File_articles_error_def_proto != nil {
 		return
 	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_api_articles_error_def_proto_rawDesc,
+			RawDescriptor: file_articles_error_def_proto_rawDesc,
 			NumEnums:      1,
 			NumMessages:   0,
 			NumExtensions: 0,
 			NumServices:   0,
 		},
-		GoTypes:           file_api_articles_error_def_proto_goTypes,
-		DependencyIndexes: file_api_articles_error_def_proto_depIdxs,
-		EnumInfos:         file_api_articles_error_def_proto_enumTypes,
+		GoTypes:           file_articles_error_def_proto_goTypes,
+		DependencyIndexes: file_articles_error_def_proto_depIdxs,
+		EnumInfos:         file_articles_error_def_proto_enumTypes,
 	}.Build()
-	File_api_articles_error_def_proto = out.File
-	file_api_articles_error_def_proto_rawDesc = nil
-	file_api_articles_error_def_proto_goTypes = nil
-	file_api_articles_error_def_proto_depIdxs = nil
+	File_articles_error_def_proto = out.File
+	file_articles_error_def_proto_rawDesc = nil
+	file_articles_error_def_proto_goTypes = nil
+	file_articles_error_def_proto_depIdxs = nil
 }
