@@ -2,7 +2,7 @@
 // versions:
 // 	protoc-gen-go v1.34.2
 // 	protoc        v4.24.3
-// source: api/helloworld/v1/error_reason.proto
+// source: helloworld/v1/error_reason.proto
 
 package v1
 
@@ -50,11 +50,11 @@ func (x ErrorReason) String() string {
 }
 
 func (ErrorReason) Descriptor() protoreflect.EnumDescriptor {
-	return file_api_helloworld_v1_error_reason_proto_enumTypes[0].Descriptor()
+	return file_helloworld_v1_error_reason_proto_enumTypes[0].Descriptor()
 }
 
 func (ErrorReason) Type() protoreflect.EnumType {
-	return &file_api_helloworld_v1_error_reason_proto_enumTypes[0]
+	return &file_helloworld_v1_error_reason_proto_enumTypes[0]
 }
 
 func (x ErrorReason) Number() protoreflect.EnumNumber {
@@ -63,43 +63,43 @@ func (x ErrorReason) Number() protoreflect.EnumNumber {
 
 // Deprecated: Use ErrorReason.Descriptor instead.
 func (ErrorReason) EnumDescriptor() ([]byte, []int) {
-	return file_api_helloworld_v1_error_reason_proto_rawDescGZIP(), []int{0}
+	return file_helloworld_v1_error_reason_proto_rawDescGZIP(), []int{0}
 }
 
-var File_api_helloworld_v1_error_reason_proto protoreflect.FileDescriptor
+var File_helloworld_v1_error_reason_proto protoreflect.FileDescriptor
 
-var file_api_helloworld_v1_error_reason_proto_rawDesc = []byte{
-	0x0a, 0x24, 0x61, 0x70, 0x69, 0x2f, 0x68, 0x65, 0x6c, 0x6c, 0x6f, 0x77, 0x6f, 0x72, 0x6c, 0x64,
-	0x2f, 0x76, 0x31, 0x2f, 0x65, 0x72, 0x72, 0x6f, 0x72, 0x5f, 0x72, 0x65, 0x61, 0x73, 0x6f, 0x6e,
-	0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x0d, 0x68, 0x65, 0x6c, 0x6c, 0x6f, 0x77, 0x6f, 0x72,
-	0x6c, 0x64, 0x2e, 0x76, 0x31, 0x2a, 0x3a, 0x0a, 0x0b, 0x45, 0x72, 0x72, 0x6f, 0x72, 0x52, 0x65,
-	0x61, 0x73, 0x6f, 0x6e, 0x12, 0x17, 0x0a, 0x13, 0x47, 0x52, 0x45, 0x45, 0x54, 0x45, 0x52, 0x5f,
-	0x55, 0x4e, 0x53, 0x50, 0x45, 0x43, 0x49, 0x46, 0x49, 0x45, 0x44, 0x10, 0x00, 0x12, 0x12, 0x0a,
-	0x0e, 0x55, 0x53, 0x45, 0x52, 0x5f, 0x4e, 0x4f, 0x54, 0x5f, 0x46, 0x4f, 0x55, 0x4e, 0x44, 0x10,
-	0x01, 0x42, 0x3e, 0x0a, 0x0d, 0x68, 0x65, 0x6c, 0x6c, 0x6f, 0x77, 0x6f, 0x72, 0x6c, 0x64, 0x2e,
-	0x76, 0x31, 0x50, 0x01, 0x5a, 0x19, 0x62, 0x6c, 0x6f, 0x67, 0x2f, 0x61, 0x70, 0x69, 0x2f, 0x68,
-	0x65, 0x6c, 0x6c, 0x6f, 0x77, 0x6f, 0x72, 0x6c, 0x64, 0x2f, 0x76, 0x31, 0x3b, 0x76, 0x31, 0xa2,
-	0x02, 0x0f, 0x41, 0x50, 0x49, 0x48, 0x65, 0x6c, 0x6c, 0x6f, 0x77, 0x6f, 0x72, 0x6c, 0x64, 0x56,
-	0x31, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
+var file_helloworld_v1_error_reason_proto_rawDesc = []byte{
+	0x0a, 0x20, 0x68, 0x65, 0x6c, 0x6c, 0x6f, 0x77, 0x6f, 0x72, 0x6c, 0x64, 0x2f, 0x76, 0x31, 0x2f,
+	0x65, 0x72, 0x72, 0x6f, 0x72, 0x5f, 0x72, 0x65, 0x61, 0x73, 0x6f, 0x6e, 0x2e, 0x70, 0x72, 0x6f,
+	0x74, 0x6f, 0x12, 0x0d, 0x68, 0x65, 0x6c, 0x6c, 0x6f, 0x77, 0x6f, 0x72, 0x6c, 0x64, 0x2e, 0x76,
+	0x31, 0x2a, 0x3a, 0x0a, 0x0b, 0x45, 0x72, 0x72, 0x6f, 0x72, 0x52, 0x65, 0x61, 0x73, 0x6f, 0x6e,
+	0x12, 0x17, 0x0a, 0x13, 0x47, 0x52, 0x45, 0x45, 0x54, 0x45, 0x52, 0x5f, 0x55, 0x4e, 0x53, 0x50,
+	0x45, 0x43, 0x49, 0x46, 0x49, 0x45, 0x44, 0x10, 0x00, 0x12, 0x12, 0x0a, 0x0e, 0x55, 0x53, 0x45,
+	0x52, 0x5f, 0x4e, 0x4f, 0x54, 0x5f, 0x46, 0x4f, 0x55, 0x4e, 0x44, 0x10, 0x01, 0x42, 0x3e, 0x0a,
+	0x0d, 0x68, 0x65, 0x6c, 0x6c, 0x6f, 0x77, 0x6f, 0x72, 0x6c, 0x64, 0x2e, 0x76, 0x31, 0x50, 0x01,
+	0x5a, 0x19, 0x62, 0x6c, 0x6f, 0x67, 0x2f, 0x61, 0x70, 0x69, 0x2f, 0x68, 0x65, 0x6c, 0x6c, 0x6f,
+	0x77, 0x6f, 0x72, 0x6c, 0x64, 0x2f, 0x76, 0x31, 0x3b, 0x76, 0x31, 0xa2, 0x02, 0x0f, 0x41, 0x50,
+	0x49, 0x48, 0x65, 0x6c, 0x6c, 0x6f, 0x77, 0x6f, 0x72, 0x6c, 0x64, 0x56, 0x31, 0x62, 0x06, 0x70,
+	0x72, 0x6f, 0x74, 0x6f, 0x33,
 }
 
 var (
-	file_api_helloworld_v1_error_reason_proto_rawDescOnce sync.Once
-	file_api_helloworld_v1_error_reason_proto_rawDescData = file_api_helloworld_v1_error_reason_proto_rawDesc
+	file_helloworld_v1_error_reason_proto_rawDescOnce sync.Once
+	file_helloworld_v1_error_reason_proto_rawDescData = file_helloworld_v1_error_reason_proto_rawDesc
 )
 
-func file_api_helloworld_v1_error_reason_proto_rawDescGZIP() []byte {
-	file_api_helloworld_v1_error_reason_proto_rawDescOnce.Do(func() {
-		file_api_helloworld_v1_error_reason_proto_rawDescData = protoimpl.X.CompressGZIP(file_api_helloworld_v1_error_reason_proto_rawDescData)
+func file_helloworld_v1_error_reason_proto_rawDescGZIP() []byte {
+	file_helloworld_v1_error_reason_proto_rawDescOnce.Do(func() {
+		file_helloworld_v1_error_reason_proto_rawDescData = protoimpl.X.CompressGZIP(file_helloworld_v1_error_reason_proto_rawDescData)
 	})
-	return file_api_helloworld_v1_error_reason_proto_rawDescData
+	return file_helloworld_v1_error_reason_proto_rawDescData
 }
 
-var file_api_helloworld_v1_error_reason_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
-var file_api_helloworld_v1_error_reason_proto_goTypes = []any{
+var file_helloworld_v1_error_reason_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
+var file_helloworld_v1_error_reason_proto_goTypes = []any{
 	(ErrorReason)(0), // 0: helloworld.v1.ErrorReason
 }
-var file_api_helloworld_v1_error_reason_proto_depIdxs = []int32{
+var file_helloworld_v1_error_reason_proto_depIdxs = []int32{
 	0, // [0:0] is the sub-list for method output_type
 	0, // [0:0] is the sub-list for method input_type
 	0, // [0:0] is the sub-list for extension type_name
@@ -107,27 +107,27 @@ var file_api_helloworld_v1_error_reason_proto_depIdxs = []int32{
 	0, // [0:0] is the sub-list for field type_name
 }
 
-func init() { file_api_helloworld_v1_error_reason_proto_init() }
-func file_api_helloworld_v1_error_reason_proto_init() {
-	if File_api_helloworld_v1_error_reason_proto != nil {
+func init() { file_helloworld_v1_error_reason_proto_init() }
+func file_helloworld_v1_error_reason_proto_init() {
+	if File_helloworld_v1_error_reason_proto != nil {
 		return
 	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_api_helloworld_v1_error_reason_proto_rawDesc,
+			RawDescriptor: file_helloworld_v1_error_reason_proto_rawDesc,
 			NumEnums:      1,
 			NumMessages:   0,
 			NumExtensions: 0,
 			NumServices:   0,
 		},
-		GoTypes:           file_api_helloworld_v1_error_reason_proto_goTypes,
-		DependencyIndexes: file_api_helloworld_v1_error_reason_proto_depIdxs,
-		EnumInfos:         file_api_helloworld_v1_error_reason_proto_enumTypes,
+		GoTypes:           file_helloworld_v1_error_reason_proto_goTypes,
+		DependencyIndexes: file_helloworld_v1_error_reason_proto_depIdxs,
+		EnumInfos:         file_helloworld_v1_error_reason_proto_enumTypes,
 	}.Build()
-	File_api_helloworld_v1_error_reason_proto = out.File
-	file_api_helloworld_v1_error_reason_proto_rawDesc = nil
-	file_api_helloworld_v1_error_reason_proto_goTypes = nil
-	file_api_helloworld_v1_error_reason_proto_depIdxs = nil
+	File_helloworld_v1_error_reason_proto = out.File
+	file_helloworld_v1_error_reason_proto_rawDesc = nil
+	file_helloworld_v1_error_reason_proto_goTypes = nil
+	file_helloworld_v1_error_reason_proto_depIdxs = nil
 }
